Move files back to their old namespace
