+++ src/HydrationMiddleware/PerformActionCalls.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Livewire\Hydration\Middleware;
+namespace Livewire\HydrationMiddleware;
 
 use Livewire\Livewire;
 use function Livewire\str;
