+++ src/HydrationMiddleware/SecureHydrationWithChecksum.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Livewire\Hydration\Middleware;
+namespace Livewire\HydrationMiddleware;
 
 use Livewire\ComponentChecksumManager;
 use Livewire\Exceptions\CorruptComponentPayloadException;
